chore: rename the template
