+++ lib-multisrc/colamanga/default/AndroidManifest.xml
